requirements.txt
fastapi
slack_bolt
SQLAlchemy
APScheduler
uvicorn
python-dotenv
pydantic-settings
asyncpg # For PostgreSQL async support
arxiv # For fetching papers from arXiv
aiohttp # For AsyncSocketModeHandler